+++ at_end2end_test/test/notify_test.dart
@@ -3,10 +3,9 @@ import 'dart:io';
 import 'package:uuid/uuid.dart';
 
 import 'package:at_client/at_client.dart';
-import 'package:at_commons/at_commons.dart';
+import 'package:at_client/src/transformer/response_transformer/notification_response_transformer.dart';
 import 'package:at_end2end_test/config/config_util.dart';
 import 'package:test/test.dart';
-import 'package:at_client/src/service/notification_service.dart';
 import 'test_utils.dart';
 
 void main() {
@@ -82,6 +81,64 @@ void main() {
     expect(notificationListJson[0]['from'], currentAtSign);
     expect(notificationListJson[0]['to'], sharedWithAtSign);
     expect(notificationListJson[0]['value'], isNotEmpty);
+  });
+
+  /// The purpose of this test is to verify the notify text with setting
+  /// shouldEncrypt parameter to true (which encrypt the notify text)
+  /// and setting shouldEncrypt to false (text message is sent as plain text).
+  group('A group of tests to verify notification text', () {
+    var notifyText = 'Hello How are you';
+    var whomToNotify = ConfigUtil.getYaml()['atSign']['secondAtSign'];
+    var inputToExpectedOutput = {
+      // Encrypt the notify text data
+      NotificationParams.forText('$notifyText', whomToNotify,
+          shouldEncrypt: true): '$whomToNotify:$notifyText',
+      // Send notify text message as plain text
+      NotificationParams.forText('$notifyText', whomToNotify,
+          shouldEncrypt: false): '$whomToNotify:$notifyText'
+    };
+    inputToExpectedOutput.forEach((input, expectedOutput) {
+      test('Setting shouldEncrypt to ${input.atKey.metadata?.isEncrypted}',
+          () async {
+        // Setting the AtClientManager instance to current atsign
+        await AtClientManager.getInstance().setCurrentAtSign(
+            currentAtSign, namespace, TestUtils.getPreference(currentAtSign));
+
+        var notificationResult = await AtClientManager.getInstance()
+            .notificationService
+            .notify(input);
+        print(notificationResult.notificationID);
+        expect(notificationResult.notificationStatusEnum,
+            NotificationStatusEnum.delivered);
+
+        // Setting the AtClientManager instance to sharedWith atsign
+        await AtClientManager.getInstance().setCurrentAtSign(sharedWithAtSign,
+            namespace, TestUtils.getPreference(sharedWithAtSign));
+        // Getting the notification from notify:list command
+        var notificationListResponse = await AtClientManager.getInstance()
+            .atClient
+            .getRemoteSecondary()
+            ?.executeCommand('notify:list\n', auth: true);
+        notificationListResponse =
+            notificationListResponse?.replaceFirst('data:', '');
+        List notificationListJSON = jsonDecode(notificationListResponse!);
+        // Filter the notification using the notification Id.
+        notificationListJSON.retainWhere(
+            (element) => element['id'] == notificationResult.notificationID);
+        expect(notificationListJSON.length, 1);
+        var response = await NotificationResponseTransformer().transform(Tuple()
+          ..one = AtNotification(
+              notificationListJSON.first['id'],
+              notificationListJSON.first['key'],
+              notificationListJSON.first['from'],
+              notificationListJSON.first['to'],
+              notificationListJSON.first['epochMillis'],
+              notificationListJSON.first['messageType'],
+              notificationListJSON.first['isEncrypted'])
+          ..two = (NotificationConfig()..shouldDecrypt = true));
+        expect(response.key, expectedOutput);
+      });
+    });
   });
 
   tearDownAll(() async {