+++ at_client/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.31
+- Enhance notify text to send text message encrypted
+- Upgrade at_persistence_secondary_server to v3.0.30
+- Upgrade at_commons version to v3.0.20 for encrypt notify text message
+- Upgrade at_lookup version to v3.0.28 for adding mutex to authenticate methods
 ## 3.0.30
 - Added bypassCache option in get method
 - Added sync conflict info to sync progress callback